requirements.txt
mcp[cli]
python-dotenv
google-genai
langchain
langchain-mcp-adapters
langchain-google-genai
google-generativeai
langgraph
pytz
requests
pint
python-pptx
pywin32
pillow
wikipedia
beautifulsoup4
